### 2020-09-04 #2639 @kthyng
Hi @skadio, thank you for your submission. Please update your paper to meet the JOSS [submission](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) requirement of:

> The paper should be between 250-1000 words.
