### 2020-09-09 #2639 @galessiorob
👋 Hi @skadio

I am in the process of assigning reviewers to this submission, but if you have any suggestions for reviewers please let me know.

In the meantime, it looks like you're only missing 3 DOIs, if you could nail those down that would be great.

Many thanks!

Basic checks:

 - [ ] There’s a general description of the software
 - [ ] The software is within the scope as research software
 - [ ] OSI-approved license: GNU General Public License (GPL)
