### 2020-09-10 #2639 @skadio
Thank you @galessiorob and thanks to our reviewer @HaoZeke!

On the basic checks:

- [ ] There’s a general description of the software
--> Indeed, a general description is available for our reviewers.

- [ ] The software is within the scope as research software
--> This library is an outcome of a research collaboration between CMU and Fidelity and will be presented at INFORMS 2020 as an invited talk in the Machine Learning & Optimization session.

- [ ]  OSI-approved license: GNU General Public License (GPL)
--> It's licensed under GNU GPL

@takojunior can you take a look at the missing DOIs and update the .bib file when you get a chance?
