+++ joss/reference.bib
@@ -4,7 +4,8 @@
   booktitle={Proceedings of the AAAI Conference on Artificial Intelligence},
   volume={33},
   pages={1495--1502},
-  year={2019}
+  year={2019},
+  doi={10.1609/aaai.v33i01.33011495}
 }
 
 @book{DBLP:series/aifta/BergmanCHH16,
@@ -26,8 +27,9 @@
 
 @misc{MPP,
   author = {Amin Hosseininasab},
-  title = {Constraint-based Sequential Pattern Mining with Decsion Diagrams},
+  title = {Constraint-based Sequential Pattern Mining with Decision Diagrams},
   year = {2019},
+  doi = {10.1609/aaai.v33i01.33011495},
   publisher = {GitHub},
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/aminhn/MPP}}
@@ -74,5 +76,6 @@
 	booktitle={Joint European Conference on Machine Learning and Knowledge Discovery in Databases},
 	pages={315--330},
 	year={2016},
+  doi={10.1007/978-3-319-46227-1_20},
 	organization={Springer}
 }
