+++ joss/reference.bib
@@ -29,7 +29,7 @@
   author = {Amin Hosseininasab},
   title = {Constraint-based Sequential Pattern Mining with Decision Diagrams},
   year = {2019},
-  doi = {},
+  doi = {10.1609/aaai.v33i01.33011495},
   publisher = {GitHub},
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/aminhn/MPP}}
