+++ joss/reference.bib
@@ -76,6 +76,6 @@
 	booktitle={Joint European Conference on Machine Learning and Knowledge Discovery in Databases},
 	pages={315--330},
 	year={2016},
-  doi={10.1007/978-3-319-46227-1_20},
+    doi={10.1007/978-3-319-46227-1_20},
 	organization={Springer}
 }
