### 2020-09-10 #2639 @takojunior
Thank you @skadio, @galessiorob, @HaoZeke!

The missing DOIs have been resolved.
