### 2020-09-10 #2662 @whedon
[REVIEW]: Seq2Pat: Sequence-to-Pattern Generation Library
**Submitting author:** @skadio (<a href="http://orcid.org/0000-0002-4672-6830">Serdar Kadioglu</a>)
**Repository:** <a href="https://github.com/fidelity/seq2pat" target ="_blank">https://github.com/fidelity/seq2pat</a>
**Version:** v1.1.1

@HaoZeke, @TimKam & @shreyasbapat, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
