### 2020-09-10 #2662 @galessiorob
@skadio the last and third to last cited reference need DOIs (not sure why whedon is not catching both of them), would you mind adding them, please? 🙏
