### 2020-09-11 #2662 @shreyasbapat
> @skadio the last and third to last cited reference need DOIs (not sure why whedon is not catching both of them), would you mind adding them, please? 🙏

I will review as soon as this is done :D
