+++ joss/reference.bib
@@ -41,7 +41,7 @@
 	year={2019},
 	journal={INFORMS Journal on Computing},
 	publisher={INFORMS},
-	note={(to appear)}
+	doi={10.1287/ijoc.2019.0937}
 }
 
 @book{wegener2000branching,
