+++ joss/reference.bib
@@ -49,6 +49,7 @@
 	author={Wegener, Ingo},
 	volume={4},
 	year={2000},
+	doi={10.1137/1.9780898719789},
 	publisher={SIAM}
 }
 
