### 2020-09-19 fidelity/seq2pat#11 @TimKam
Review Issue, @TimKam
Dear all,

let us use this issue to discuss my review (you can branch off smaller issues if you prefer).
You find the initial review below.

Best,
Timotheus

------
The contribution introduces a Python-based sequence-to-pattern generation library that is based on/re-uses [recent research results](https://aaai.org/ojs/index.php/AAAI/article/view/3962) in the domain.

All in all, the contribution seems to be on the way to a good state, with the following limitations:

0. Contribution and authorship: The main author and generally 3 of 5 authors have made substantial contributions to the software. The most senior "not contributing author" seems to be a leading expert on the topic. All in all, this looks like a reasonable list of authors (given common academic practice), even though not all authors have contributed code/docs. (No action needed)

1. Substantial scholarly effort: in essence, the library is a wrapper around research code that has been written for the corresponding [AAAI 2019 paper](https://aaai.org/ojs/index.php/AAAI/article/view/3962). The authors should be more clear about this fact. (The paper is prominently linked; for sure, the authors did not aim to disguise this fact.) This does not mean that the contribution is not valuable, but it certainly has implications on the _additional effort_ that was necessary to create the library. Here, I'd like to get the perspective of an editor.

2. Automated tests: It would be good to configure a CI for the repository. Even for review purposes this would have been neat. No documentation on how to run the tests exists. Navigating to ``tests`` and running ``pytest test_seq2pat.py`` executes 40 tests, at least some of which have multiple assertions. All tests seem to be managed in one file (1244 LOC), which is not according to good practices. At least unit tests should roughly reflect the project structure.

3. Examples and API documentation: Very simple examples are provided and the API is documented.

	3.1. However, the examples are largely abstract (the concepts of "items" and "price" are used in the examples, but that's about it) and do not illustrate the practical problems the library can potentially solve. This is very unfortunate, because for a non-expert reader it is very hard to understand the practical benefit of the library. I strongly recommend to add examples that relate to a potential application domain, even if these examples are merely toy examples.

	Nit-picks:

	3.2. In the docs, it seems a bit odd to me that the API doc has the heading "Seq2Pat _Public_ API". _Public_ vs. _Private_ 1) does not exist in Python and 2) is in the context of APIs more commonly used in commercial software/software-as-a-service scenarios.

	3.3. Currently, the repository contains the sources of the docs _and_ the Sphinx build result. Using a free-of-charge service like [Read the Docs](https://readthedocs.org/) for automatically pulling the sources from GitHub and building them is more elegant.

4. Community guidelines: Community guidelines for contributing do not exist.

5. Statement of need: a statement of need exists in the _paper_, not in the docs. The paper could benefit from a less abstract perspective on the practical usefulness of the library (same as with the examples, see above). Considering that the software is co-developed by some sort of investment firm, providing a practical motivation should be possible.

6. Summary: see above, an introduction for non-experts is missing.

7. State of the field: the authors provide an overview of the state of the field, but do not motivate why they create a new library instead of improving existing libraries.

8. The authors make good use of references. However, in the introduction, some references that point to high-level overviews, applied research papers and/or surveys would be nice.
