### 2020-09-19 #2662 @TimKam
I have submitted my initial evaluation here: https://github.com/fidelity/seq2pat/issues/11.

Regarding the following evaluation result, I'd like to get the perspective of an editor (@galessiorob):
In essence, the library is a wrapper around research code that has been written for the corresponding [AAAI 2019 paper](https://aaai.org/ojs/index.php/AAAI/article/view/3962). The authors should be more clear about this fact. (The paper is prominently linked; for sure, the authors did not aim to disguise this fact.) This does not mean that the contribution is not valuable, but it certainly has implications on the _additional effort_ that was necessary to create the library.

Limitations: I don't have sufficient knowledge of sequence-to-pattern generation and multi-valued decision diagrams to judge aspects of the contribution that go beyond the software engineering perspective (I merely recognize some of the authors cited in the foundational AAAI '19 paper from "my" community).
