### 2020-09-21 fidelity/seq2pat#11 @skadio
@TimKam thank you very much for your thorough review! Let's follow up with our updates/comments.
