### 2020-09-22 #2662 @galessiorob
Thanks for raising this question @TimKam! Based on our existing lineaments while brief, this submission is valid, as for the "extensibility/adaptability" for code that builds on existing work I _think_ this paper also passes but let me get an editor in chief's opinion.
