### 2020-09-29 #2662 @Kevin-Mattheus-Moerman
@skadio thanks for your submission to JOSS. Several reviewers raised concerns in terms of this work potentially being out of scope. The handling editor, @galessiorob, consulted the JOSS editorial board who has now also reviewed and discussed this work in detail. Unfortunately the editorial board agreed that this work is out of scope for JOSS as it stands. In particular it is deemed too minor in terms of its functionality (see [our guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)). This conclusion does not mean the software is not useful or of a low quality, it merely means that in terms of required amount of scholarly content this work is out of scope for JOSS.

I will reject this submission at this point. However, if you continue to develop and expand this work, we would be happy to reconsider it as a re-submission in the future. Alternatively you could archive a copy of this work on a service like ZENODO to obtain a DOI which users can cite.

Dr Kevin Moerman
_AEiC JOSS_
