### 2020-10-23 #2662 @skadio
@Kevin-Mattheus-Moerman @galessiorob

I would like to thank you again for your time and service! I have read the guidelines as well as your response about the Scope and kindly disagree, especially when compared to other submissions that I have been part of. Leaving this aside, we already received an initial evaluation from @TimKam. Thank you very much @TimKam, we incorporated some of the input already and your pointers will help us improve over the current work.

In return for the time of the JOSS reviewers here, I agreed to provide evaluations on a number of submissions and encourage our main co-authors here to do the same.

This was our first submission to JOSS, thank you again for your feedback!

Best,
Serdar
